move xmltoTreeElement to TreeUtils
